+++ planetiler-custommap/src/main/java/com/onthegomap/planetiler/custommap/configschema/MergeOverlappingPolygons.java
@@ -3,5 +3,5 @@ package com.onthegomap.planetiler.custommap.configschema;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record MergeOverlappingPolygons(
-  @JsonProperty("min_area") int minArea
+  @JsonProperty("min_area") double minArea
 ) {}
